Stop writing metadata to report when saving
When saving a report using the report plugin it was saving the current
version of the resource's metadata to the updated resource. This would
result in the incorrect last modified date being presented.
When fetching the resource we now use Resource.getSource() which
doesn't include the metadata, whereas Resource.get() did