3996 // Studio rows should include index in Key because they can have same      self
Sometimes different rows in Studio table can have same self. (Example,
"Cell Types" Studio in neuroscience organization in staging env). In
these cases, we still need to have a unique React.Key for each row.
However, self should still be used to fetch resources for download.

That's why a row in studio needs two keys:

tableKey: to help react render each row in table correctly. This is
unique for every row
localStorageKey: usually the value of `self.value`. This is used to
retrieve a row to display it in the resources panel or to fetch it for
downloading it through the data panel.

Signed-off-by: Dinika Saxena <dinikasaxenas@gmail.com>