Download of resources with tags in their ids should work
When storing these resources and their distributions to the localstorage
the tag should be stored as well. Otherwise the `key` (required to
add/remove items from data panel) is inconsistent, leading to bugs
(example, adding the resource works, but removing them from the data
panel does not, because the id is not the same).

Signed-off-by: Dinika Saxena <dinikasaxenas@gmail.com>